[FEATURE] add read and update for transtype

+++ src/main/java/com/enigma/wmb_api/dto/request/SearchMenuRequest.java
@@ -13,6 +13,6 @@ public class SearchMenuRequest {
     private String sortBy;
     private String direction;
     private String name;
-    private Float minPrice;
-    private Float maxPrice;
+    private Integer minPrice;
+    private Integer maxPrice;
 }
